refactor(api): simplify branch api structure
