+++ package.json
@@ -92,7 +92,6 @@
     "@orpc/client": "^1.14.0",
     "@orpc/contract": "1.14.0",
     "@orpc/openapi": "1.14.0",
-    "@orpc/openapi-client": "1.14.0",
     "@orpc/server": "^1.14.0",
     "@orpc/tanstack-query": "^1.14.0",
     "@orpc/zod": "1.14.0",
